:truck: Move auth hook to its own folder
